pyproject.toml
[tool.poetry]
name = "ai-agents"
version = "0.1.0"
description = ""
authors = ["Santo K Thomas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
eval-type-backport = "^0.2.0"
griffe = "^1.5.1"
httpx = "^0.28.1"
logfire-api = "^2.7.1"
pydantic = "^2.10.3"
google-auth = "^2.37.0"
requests = "^2.32.3"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
instructor = "^1.7.0"
google-cloud-aiplatform = "^1.75.0"
jsonref = "^1.1.0"
pypdf2 = "^3.0.1"
google-generativeai = "^0.8.3"
google-genai = "^0.3.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
